pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bridge_detector_api"
version = "0.1.2"
description = "API for detecting bridges on images"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
  { name = "Marat Saitov", email = "maratsajtov@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Framework :: Flask",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Software Development :: Libraries :: Application Frameworks"
]

dependencies = [
    "Flask==3.1.1",
    "numpy==2.3.1",
    "opencv-python==4.11.0.86",
    "pillow==11.2.1",
    "ultralytics==8.3.160",
    "torch==2.7.1",
    "torchvision==0.22.1",
    "Werkzeug==3.1.3",
    "click==8.2.1",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.2"
]

[project.optional-dependencies]
test = [
    "pytest==8.4.1",
    "pytest-cov==6.2.1",
    "pytest-flask==1.3.0",
    "pytest-mock==3.14.1",
    "coverage==7.9.1",
    "requests==2.32.4"
]

dev = [
    "flake8==7.3.0",
    "mypy==1.5.1",
    "pycodestyle==2.14.0",
    "pyflakes==3.4.0",
    "mccabe==0.7.0",
    "tox==4.27.0",
    "virtualenv==20.31.2",
    "build==1.2.2.post1",
    "pyproject-hooks==1.2.0",
    "platformdirs==4.3.8",
    "filelock==3.18.0",
    "distlib==0.3.9"
]

doc = [
    "Sphinx==8.2.3",
    "alabaster==1.0.0",
    "babel==2.17.0",
    "docutils==0.21.2",
    "imagesize==1.4.1",
    "snowballstemmer==3.0.1",
    "sphinxcontrib-applehelp==2.0.0",
    "sphinxcontrib-devhelp==2.0.0",
    "sphinxcontrib-htmlhelp==2.1.0",
    "sphinxcontrib-jsmath==1.0.1",
    "sphinxcontrib-qthelp==2.0.0",
    "sphinxcontrib-serializinghtml==2.0.0",
    "Pygments==2.19.2",
    "pyparsing==3.2.3",
    "roman-numerals-py==3.1.0"
]

[tool.setuptools.packages.find]
include = ["bridge_detector_api*"]

[tool.setuptools.package-data]
bridge_detector_api = ["assets/*.pt"]

[project.urls]
Homepage = "https://example.com"
Repository = "https://github.com/saitovmarat/bridge-detector-api"
Documentation = "https://example.com/docs"

[project.scripts]
bridge_detector_api = "bridge_detector_api.cli:cli"
